Divide Form component into Title, Score, Body, Reviewer components
